requirements.txt
requests==2.21.0
python-telegram-bot==20.0a2
psycopg2==2.9.3